converted signup component to use hooks and created a signup modal successfully signing up a user
