+++ frontend/src/App.js
@@ -10,7 +10,7 @@ import Signup from './components/Signup';
 import { loadToken } from './store/reducers/auth';
 import ProtectedRoute from './components/ProtectedRoute';
 import Demo from './components/Demo';
-import LandingPageHooks from './components/LandingPageHooks';
+
 
 
 class App extends Component {
@@ -33,7 +33,7 @@ class App extends Component {
 		return (
 			<BrowserRouter>
 				<Switch>
-					<Route exact path='/' component={LandingPageHooks} />
+					<Route exact path='/' component={LandingPage} />
 					<ProtectedRoute
 						isLoggedIn={this.props.token}
 						path='/browse'
